feat: Integrate FileExists function to skip downloading existing files
- Before downloading, the function now checks if the file already exists.
- If the file exists, the function skips the download to prevent redundancy.

+++ wgetApp/asyncMIrror.go
@@ -74,6 +74,10 @@ func (app *WgetApp) asyncMirror(outputFile, urls, direc string) error {
 		}
 	}
 
+	if wgetutils.FileExists(outputFile) {
+		return nil
+	}
+
 	var out *os.File
 	out, err = os.Create(outputFile)
 	if err != nil {
